moved installer of refind and grub into a bootloaders.py file.
