+++ scripts/archinstall/arch/__init__.py
@@ -5,3 +5,4 @@ from . import services
 from . import tools
 from . import exceptions
 from . import metapkg
+from . import bootloaders
